pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "Gradients-on-Demand"
version = "0.0.3"
description = "Validator and Miner for Gradients on Demand"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "Apache-2.0" }
authors = [
    { name = "Gradients.io", email = "gradientsio@outlook.com" },
    { name = "namoray", email = "namoray@rayonlabs.ai" },
]
dependencies = [
    "opentelemetry-distro==0.47b0",
    "opentelemetry-exporter-otlp-proto-http==1.26.0",
    "opentelemetry-instrumentation-asgi==0.47b0",
    "opentelemetry-instrumentation-asyncio==0.47b0",
    "opentelemetry-instrumentation-asyncpg==0.47b0",
    "opentelemetry-instrumentation-fastapi==0.47b0",
    "opentelemetry-instrumentation-httpx==0.47b0",
    "opentelemetry-instrumentation-logging==0.47b0",
    "opentelemetry-instrumentation-redis==0.47b0",
    "opentelemetry-instrumentation-system-metrics==0.47b0",
    "scalar-fastapi==1.0.3",
    "fastapi>=0.110.3,<0.113.0",
    "uvicorn>=0.30.5,<0.32.0",
    "fiber @ git+https://github.com/rayonlabs/fiber.git@2.4.0",
    "asyncpg==0.29.0",
    "httpx==0.27.0",
    "loguru==0.7.2",
    "python-dotenv==1.0.1",
    "redis==5.0.8",
    "scipy>=1.10.0",
    "numpy>=2.0.0",
    "datasets>=2.16.0,<3.0.0",
    "docker==7.1.0",
    "tenacity==9.0.0",
    "minio==7.2.10",
    "sqlalchemy==2.0.36",
    "transformers==4.46.2",
    "pandas==2.2.3",
    "tiktoken==0.8.0",
    "sentencepiece==0.2.0",
    "peft",
    "Pillow==11.1.0",
    "textstat==0.7.7",
    "langcheck==0.9.0",
    "detoxify",
    "gitpython",
    "nvidia-ml-py3",
    "aiofiles",
    "astor",
    "cryptography",
]

[project.optional-dependencies]
dev = [
    "asgi-lifespan==2.1.0",
    "freezegun>=1.5.1",
    "ipython==8.28.0",
    "pre-commit==4.0.1",
    "pytest-asyncio==0.21.1",
    "pytest-cov==5.0.0",
    "pytest-random-order>=1.1.1",
    "pytest~=8.1.0",
    "ruff>=0.6.8",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.hatch.build]
include = ["src/**/*.py"]

[tool.hatch.build.targets.sdist]
include = ["fiber"]

[tool.ruff]
line-length = 130

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = ["E402"]

[tool.ruff.lint.isort]
lines-after-imports = 2
force-single-line = true

[tool.hatch.metadata]
allow-direct-references = true

[tool.typos.default]
extend-ignore-identifiers-re = [
    # ignore a single line by putting this on the end.
    "(?Rm)^.*(#|//)\\s*spellcheck:ignore$",

    # ignore a block by surrounding it with spellcheck:on and spellcheck:off
    "(?s)(#|//)\\s*spellcheck:off.*?\\n\\s*(#|//)\\s*spellcheck:on",

    # ss58 hotkeys
    "[a-zA-Z0-9]{48}",
]

[tool.pytest.ini_options]
addopts = "--random-order --cov=src --cov-report=term-missing --cov-report=xml"
asyncio_mode = "auto"


# Missing pyright from here
